==> _archives/_archive_4/scraped_data/2_tmdb_data.ipynb <==
# import pandas as pd
# import requests
# import os
# from tqdm import tqdm

# def get_tmdb_id(imdb_id, api_key):
#     url = f'https://api.themoviedb.org/3/find/{imdb_id}?api_key={api_key}&external_source=imdb_id'
#     response = requests.get(url)
#     data = response.json()
#     if data['movie_results']:
#         return data['movie_results'][0]['id']
#     else:
#         return None

# # Get current working directory
# current_working_dir = os.getcwd()

# # Construct the full path to your file
# file_path = os.path.join(current_working_dir, "data", "oscar_movies.csv")

# # load data
# df = pd.read_csv(file_path)

# # replace with your actual API key
# api_key = 'a'

# # apply the function to get the tmdb_id with a progress bar
# tqdm.pandas()
# df['tmdb_id'] = df['imdb_id'].progress_apply(lambda x: get_tmdb_id(x, api_key))

# # saving the dataframe to a new csv file
# df.to_csv('./data/oscar_movies_with_tmdb_id.csv', index=False)


# Only retrieve data for Movies, and in order to do so filter by remove null resuls for TMDb
import pandas as pd
import requests
import os
from tqdm import tqdm
import numpy as np

tqdm.pandas()

def get_tmdb_id(imdb_id, api_key):
    url = f'https://api.themoviedb.org/3/find/{imdb_id}?api_key={api_key}&external_source=imdb_id'
    response = requests.get(url)
    data = response.json()
    if data['movie_results']:
        return data['movie_results'][0]['id']

    
    # elif data['tv_results']:
    #     return data['tv_results'][0]['id']
    else:
        return None

# Get current working directory
current_working_dir = os.getcwd()

# Construct the full path to your file
file_path = os.path.join(current_working_dir, "data", "oscar_movies.csv")

# load data
df = pd.read_csv(file_path)

# replace with your actual API key
api_key = 'a'

# apply the function to get the tmdb_id
df['tmdb_id'] = df['imdb_id'].progress_apply(lambda x: get_tmdb_id(x, api_key))

# replace None with np.nan and convert tmdb_id to integer, then replace np.nan back to None
df['tmdb_id'].replace({None: np.nan}, inplace=True)
df['tmdb_id'] = df['tmdb_id'].astype(float).astype('Int64')
df['tmdb_id'].replace({np.nan: None}, inplace=True)

# erasing null values to ensure data integrity
df = df.dropna(subset=['tmdb_id'])

# saving the dataframe to a new csv file
df.to_csv('./data/oscar_movies_with_tmdb_id.csv', index=False)
