==> _archives/_archive_4/scraped_data/3_movie_to_person.ipynb <==
# import pandas as pd
# import time
# import requests
# import os
# from tqdm import tqdm

# def get_movie_credits(tmdb_id, api_key):
#     url = f'https://api.themoviedb.org/3/movie/{tmdb_id}/credits?api_key={api_key}'
#     response = requests.get(url)
#     data = response.json()
#     return data

# # Get current working directory
# current_working_dir = os.getcwd()

# # Construct the full path to your file
# file_path = os.path.join(current_working_dir, "data", "oscar_movies_with_tmdb_id.csv")

# # load data
# df = pd.read_csv(file_path)

# # replace with your actual API key
# api_key = 'a'

# # Initialize an empty DataFrame to hold our movie credits data
# df_credits = pd.DataFrame()

# for tmdb_id in tqdm(df['tmdb_id']):
#     credits_data = get_movie_credits(int(tmdb_id), api_key)

#     for cast in credits_data['cast']:
#         row = {
#             'movie_id': tmdb_id,
#             'person_id': cast['id'],
#             'person_name': cast['name'],
#             'role': 'Cast',
#             'character_or_job': cast['character'],
#             'credit_id': cast['credit_id'],
#             'uncredited': 'uncredited' in cast['character'].lower()
#         }

#         df_credits = pd.concat([df_credits, pd.DataFrame([row])], ignore_index=True)

#     for crew in credits_data['crew']:
#         if crew['job'].lower() in ['casting', 'director']:
#             row = {
#                 'movie_id': tmdb_id,
#                 'person_id': crew['id'],
#                 'person_name': crew['name'],
#                 'role': crew['job'],
#                 'character_or_job': crew['job'],
#                 'credit_id': crew['credit_id'],
#                 'uncredited': False
#             }

#             df_credits = pd.concat([df_credits, pd.DataFrame([row])], ignore_index=True)

#     time.sleep(0.3)  # delay to avoid hitting rate limit

# # save the credits dataframe to a csv file
# df_credits.to_csv('./data/movie_credits.csv', index=False)


# import pandas as pd
# import time
# import requests
# import os
# from tqdm import tqdm

# def get_movie_data(tmdb_id, api_key):
#     url = f'https://api.themoviedb.org/3/movie/{tmdb_id}?api_key={api_key}'
#     response = requests.get(url)
#     data = response.json()
#     return data

# def get_movie_credits(tmdb_id, api_key):
#     url = f'https://api.themoviedb.org/3/movie/{tmdb_id}/credits?api_key={api_key}'
#     response = requests.get(url)
#     data = response.json()
#     return data

# # Get current working directory
# current_working_dir = os.getcwd()

# # Construct the full path to your file
# input_file_path = os.path.join(current_working_dir, "data", "oscar_movies_with_tmdb_id.csv")
# output_file_path = os.path.join(current_working_dir, "data", "movie_credits.csv")

# # load data
# df = pd.read_csv(input_file_path)

# # replace with your actual API key
# api_key = 'a'

# # Try to load the existing credits data, if it fails initialize an empty DataFrame
# if os.path.exists(output_file_path):
#     df_credits = pd.read_csv(output_file_path)
# else:
#     df_credits = pd.DataFrame()

# for tmdb_id in tqdm(df['tmdb_id']):
#     movie_data = get_movie_data(int(tmdb_id), api_key)
#     imdb_id = movie_data['imdb_id']
#     credits_data = get_movie_credits(int(tmdb_id), api_key)

#     for cast in credits_data['cast']:
#         row = {
#             'movie_id': int(tmdb_id),
#             'imdb_id': imdb_id,
#             'person_id': cast['id'],
#             'person_name': cast['name'],
#             'role': 'Cast',
#             'character_or_job': cast['character'],
#             'credit_id': cast['credit_id'],
#             'uncredited': 'uncredited' in cast['character'].lower()
#         }

#         df_credits = pd.concat([df_credits, pd.DataFrame([row])], ignore_index=True)

#     for crew in credits_data['crew']:
#         if crew['job'].lower() in ['casting', 'director']:
#             row = {
#                 'movie_id': int(tmdb_id),
#                 'imdb_id': imdb_id,
#                 'person_id': crew['id'],
#                 'person_name': crew['name'],
#                 'role': crew['job'],
#                 'character_or_job': crew['job'],
#                 'credit_id': crew['credit_id'],
#                 'uncredited': False
#             }

#             df_credits = pd.concat([df_credits, pd.DataFrame([row])], ignore_index=True)

#     # Save the updated credits data
#     df_credits.to_csv(output_file_path, index=False)

#     time.sleep(0.3)  # delay to avoid hitting rate limit


import pandas as pd
import time
import requests
import os
from tqdm import tqdm

def get_movie_data(tmdb_id, api_key):
    url = f'https://api.themoviedb.org/3/movie/{tmdb_id}?api_key={api_key}'
    response = requests.get(url)
    data = response.json()
    return data

def get_movie_credits(tmdb_id, api_key):
    url = f'https://api.themoviedb.org/3/movie/{tmdb_id}/credits?api_key={api_key}'
    response = requests.get(url)
    data = response.json()
    return data

# Get current working directory
current_working_dir = os.getcwd()

# Construct the full path to your file
input_file_path = os.path.join(current_working_dir, "data", "oscar_movies_with_tmdb_id.csv")
output_file_path = os.path.join(current_working_dir, "data", "movie_to_persons.csv")

# load data
df = pd.read_csv(input_file_path)

# replace with your actual API key
api_key = 'a'

# Initialize the dataframe with the same columns
df_credits = pd.DataFrame(columns=['movie_id', 'imdb_id', 'person_id', 'person_name', 'department', 'role', 'character_or_job', 'credit_id'])

for tmdb_id in tqdm(df['tmdb_id']):
    movie_data = get_movie_data(int(tmdb_id), api_key)
    imdb_id = movie_data['imdb_id']
    credits_data = get_movie_credits(int(tmdb_id), api_key)

    for cast in credits_data['cast']:
        role = 'Cast_Uncredited' if 'uncredited' in cast['character'].lower() else 'Cast'
        row = {
            'movie_id': int(tmdb_id),
            'imdb_id': imdb_id,
            'person_id': cast['id'],
            'person_name': cast['name'],
            'department': 'Acting',  # Department for actors/actresses can be set to 'Acting'
            'role': role,
            'character_or_job': cast['character'],
            'credit_id': cast['credit_id']
        }

        df_credits = pd.concat([df_credits, pd.DataFrame([row])], ignore_index=True)
        # df_credits = df_credits.append(row, ignore_index=True)

    for crew in credits_data['crew']:
        if crew['department'].lower() in ['directing', 'production', 'writing']:
            row = {
                'movie_id': int(tmdb_id),
                'imdb_id': imdb_id,
                'person_id': crew['id'],
                'person_name': crew['name'],
                'department': crew['department'],  # This will set the 'department' to the respective department of the crew member
                'role': crew['job'],
                'character_or_job': crew['job'],
                'credit_id': crew['credit_id']
            }

            # df_credits = df_credits.append(row, ignore_index=True)
            df_credits = pd.concat([df_credits, pd.DataFrame([row])], ignore_index=True)

# Save the updated credits data
df_credits.to_csv(output_file_path, index=False)

# time.sleep(0.3)  # delay to avoid hitting rate limit


 # https://api.themoviedb.org/3/movie/667538?api_key=4ea37a8158ca566deb7455166dc1edda&append_to_response=credits

import pandas as pd

def adjust_department(row):
    if 'casting' in row['role'].lower() and row['department'] == 'Production':
        return 'Production_Casting'
    elif row['department'] == 'Acting' and row['role'] == 'Cast':
        return 'Acting'
    elif row['department'] == 'Acting' and row['role'] == 'Cast_Uncredited':
        return 'Acting_Uncredited'
    else:
        return row['department']

# Load the data
movie_to_persons = pd.read_csv('./data/movie_to_persons.csv')

# Apply the function to each row in the DataFrame
movie_to_persons['department'] = movie_to_persons.apply(adjust_department, axis=1)

# Write the DataFrame back to the CSV
movie_to_persons.to_csv('./data/movie_to_persons.csv', index=False)
