==> _archives/_archive_4/scraped_data/4_movie_detail.ipynb <==
# TODO: Based on ./data/oscar_movies_with_tmdb_id.csv, retrieve movies data from tmdb_id

# import time
# import pandas as pd
# import requests
# import os
# from tqdm import tqdm

# def get_movie_data(tmdb_id, api_key):
#     url = f'https://api.themoviedb.org/3/movie/{tmdb_id}?api_key={api_key}'
#     response = requests.get(url)
#     data = response.json()
#     return data

# # Get current working directory
# current_working_dir = os.getcwd()

# # Construct the full path to your file
# file_path = os.path.join(current_working_dir, "data", "oscar_movies_with_tmdb_id.csv")

# # load data
# df = pd.read_csv(file_path)

# # replace with your actual API key
# api_key = 'a'

# # Create a new DataFrame to store movie data
# df_movie = pd.DataFrame()

# for tmdb_id in tqdm(df['tmdb_id']):
#     data = get_movie_data(int(tmdb_id), api_key)

#     # extract the needed information from the data
#     row = {
#         'id': data['id'], 
#         'imdb_id': data['imdb_id'], 
#         # 'original_title': data['original_title'], 
#         'popularity': data['popularity'], 
#         'release_date': data['release_date'], 
#         'revenue': data['revenue'], 
#         'title': data['title'],
#     }

#     for genre in data['genres']:
#         genre_column = 'genre_' + genre['name']
#         if genre_column not in df_movie.columns:
#             df_movie[genre_column] = 0
#         row[genre_column] = 1

#     # Append the row to the DataFrame
#     df_movie = pd.concat([df_movie, pd.DataFrame([row])], ignore_index=True)
#     # print(df_movie.tail())
#     time.sleep(0.3)  # delay to avoid hitting rate limit

# # saving the dataframe to a new csv file
# df_movie.to_csv('./data/movie_data.csv', index=False)


# # Attempt at generating a palcehodler csv file to base further data concatenation. this does not work
# # Previous method of rebuilding dataframe every time allows for a more flexible way of data cleaning

# import time
# import pandas as pd
# import requests
# import os
# from tqdm import tqdm

# def get_movie_data(tmdb_id, api_key):
#     url = f'https://api.themoviedb.org/3/movie/{tmdb_id}?api_key={api_key}'
#     response = requests.get(url)
#     data = response.json()
#     return data

# # Get current working directory
# current_working_dir = os.getcwd()

# # Construct the full path to your file
# file_path = os.path.join(current_working_dir, "data", "oscar_movies_with_tmdb_id.csv")

# # load data
# df = pd.read_csv(file_path)

# # replace with your actual API key
# api_key = 'a'

# # Initialize an empty DataFrame to hold our movie data
# df_movie = pd.DataFrame()

# # Check if the movie_data.csv file exists, if it does, read it
# movie_data_file = os.path.join(current_working_dir, "data", "movie_data.csv")
# if os.path.exists(movie_data_file):
#     df_movie = pd.read_csv(movie_data_file)

# for tmdb_id in tqdm(df['tmdb_id']):
#     if tmdb_id not in df_movie['id'].values:
#         data = get_movie_data(int(tmdb_id), api_key)

#         # extract the needed information from the data
#         row = {
#             'id': data['id'], 
#             'imdb_id': data['imdb_id'], 
#             'original_title': data['original_title'], 
#             'popularity': data['popularity'], 
#             'release_date': data['release_date'], 
#             'revenue': data['revenue'], 
#             'title': data['title'],
#         }

#         for genre in data['genres']:
#             genre_column = 'genre_' + genre['name']
#             if genre_column not in df_movie.columns:
#                 df_movie[genre_column] = 0
#             row[genre_column] = 1

#         # Append the row to the DataFrame
#         df_movie = pd.concat([df_movie, pd.DataFrame([row])], ignore_index=True)

#         # save the updated dataframe to the csv file
#         df_movie.to_csv(movie_data_file, index=False)

#         time.sleep(0.3)  # delay to avoid hitting rate limit


# import time
# import pandas as pd
# import requests
# import os
# from tqdm import tqdm

# def get_movie_data(tmdb_id, api_key):
#     url = f'https://api.themoviedb.org/3/movie/{tmdb_id}?api_key={api_key}'
#     response = requests.get(url)
#     data = response.json()
#     return data

# # Get current working directory
# current_working_dir = os.getcwd()

# # Construct the full path to your file
# file_path = os.path.join(current_working_dir, "data", "oscar_movies_with_tmdb_id.csv")

# # load data
# df = pd.read_csv(file_path)

# # replace with your actual API key
# api_key = 'a'

# # Create a new DataFrame to store movie data
# df_movie = pd.DataFrame()

# for tmdb_id in tqdm(df['tmdb_id']):
#     data = get_movie_data(int(tmdb_id), api_key)

#     # extract the needed information from the data
#     row = {
#         'id': data['id'], 
#         'imdb_id': data['imdb_id'], 
#         'popularity': data['popularity'], 
#         'release_date': data['release_date'], 
#         'revenue': data['revenue'], 
#         'title': data['title'],
#     }

#     for genre in data['genres']:
#         genre_column = 'genre_' + genre['name']
#         if genre_column not in df_movxie.columns:
#             df_movie[genre_column] = 0
#         row[genre_column] = 1

#     # Append the row to the DataFrame
#     df_movie = pd.concat([df_movie, pd.DataFrame([row])], ignore_index=True)
    
#     # Save the DataFrame to a csv file
#     df_movie.to_csv('./data/movie_data.csv', index=False)
    
#     time.sleep(0.3)  # delay to avoid hitting rate limit


# https://api.themoviedb.org/3/movie/667538?api_key=4ea37a8158ca566deb7455166dc1edda

import time
import pandas as pd
import requests
import os
from tqdm import tqdm

def get_movie_data(tmdb_id, api_key):
    url = f'https://api.themoviedb.org/3/movie/{tmdb_id}?api_key={api_key}'
    response = requests.get(url)
    data = response.json()
    return data

# Get current working directory
current_working_dir = os.getcwd()

# Construct the full path to your file
file_path = os.path.join(current_working_dir, "data", "oscar_movies_with_tmdb_id.csv")

# load data
df = pd.read_csv(file_path)

# replace with your actual API key
api_key = 'a'

# Create a new DataFrame to store movie data
df_movie = pd.DataFrame()

for tmdb_id in tqdm(df['tmdb_id']):
    data = get_movie_data(int(tmdb_id), api_key)

    # extract the needed information from the data
    row = {
        'id': data['id'], 
        'imdb_id': data['imdb_id'], 
        'popularity': round(data['popularity']), 
        'vote_average': data['vote_average'], 
        'vote_count': round(data['vote_count']), 
        'release_date': data['release_date'], 
        'revenue': round(data['revenue']), 
        'title': data['title'],
        'language':data['original_language']
    }

    for genre in data['genres']:
        genre_column = 'genre_' + genre['name']
        if genre_column not in df_movie.columns:
            df_movie[genre_column] = 0
        row[genre_column] = 1

    # Append the row to the DataFrame
    df_movie = pd.concat([df_movie, pd.DataFrame([row])], ignore_index=True)

    # # Save the DataFrame to a csv file
    # df_movie.to_csv('./data/movie_data.csv', index=False)
    
    # time.sleep(0.3)  # delay to avoid hitting rate limit

# # reorder columns to group all genre columns at the end
# genre_columns = ['genre_Action', 'genre_Adventure', 'genre_Animation']  # replace with your actual list of genre columns
# other_columns = [col for col in df_movie.columns if col not in genre_columns]
# new_order = other_columns + genre_columns
# df_movie = df_movie.reindex(columns=new_order)

# save the reordered DataFrame
df_movie.to_csv('./data/movie_data.csv', index=False)


import pandas as pd

# Read the data
df = pd.read_csv('./data/movie_data.csv')

# Update column names
df.columns = df.columns.str.lower().str.replace(' ', '')

# Overwrite the original CSV file with the updated DataFrame
df.to_csv('./data/movie_data.csv', index=False)
