==> _archives/_archive_4/scraped_data/8_data_cleanup.ipynb <==
# Data Retrieval, Cleanup, and Loading

# TODO: For all data retrieval, ensure column names are lowercased and empty spaces are removed

# import pandas as pd

# # import the movie genre,
# movie_data = pd.read_csv('./data/movie_data.csv')
# # TODO: set 'id' as primary key
# # TODO: set 'id', 'title', 'imdb_id' as indexes separately
# movie_data.set_index('id', inplace=True)

# # import the movie_to_persons.csv
# # TODO: set 'movie_id, person_id' as primary key
# # TODO: set 'movie_id', 'person_id', 'imdb_id', 'department' as indexes separately
# # TODO: only bring in the columns 'movie_id', 'person_id', 'imdb_id', 'department' 
# movie_to_persons = pd.read_csv('./data/movie_to_persons.csv')
# movie_to_persons.set_index(['movie_id', 'person_id', 'role'], inplace=True)

# # import the person_data.csv
# # TODO: set 'id' as primary key
# # TODO: set 'id', 'imdb_id', 'name' as indexes separately
# # TODO: bring in all columns
# person_data = pd.read_csv('./data/person_data.csv')
# person_data.set_index('id', inplace=True)

# # import the person_name_race.csv
# # TODO: set 'id' as primary key
# # TODO: set 'id', 'name', 'name_race' as indexes separately
# # TODO: bring in all columns
# person_race = pd.read_csv('./data/person_name_race.csv')
# person_race.set_index('id', inplace=True)

# # import the person_name_race.csv
# # TODO: set 'id' as primary key
# # TODO: set 'id', 'name', 'image_race' as indexes separately
# # TODO: bring in 'id', 'name', 'image_race' columns
# person_race = pd.read_csv('./data/person_name_race.csv')
# person_race.set_index('id', inplace=True)


import pandas as pd

# import the movie genre,
movie_data = pd.read_csv('./data/movie_data.csv')
movie_data.set_index('id', inplace=True) # set 'id' as primary key
movie_data.index.name = 'id'  # Ensuring the index has a name
print(movie_data.head())  # Print the first few rows to check

# import the movie_to_persons.csv
# only bring in the columns 'movie_id', 'person_id', 'imdb_id', 'department' 
movie_to_persons = pd.read_csv('./data/movie_to_persons.csv', usecols=['movie_id', 'person_id', 'imdb_id', 'department'])
movie_to_persons.set_index(['movie_id', 'person_id', 'imdb_id', 'department'], inplace=True) # set 'movie_id, person_id, imdb_id, department' as primary key
print(movie_to_persons.head())  # Print the first few rows to check

# import the person_data.csv
# bring in all columns
person_data = pd.read_csv('./data/person_data.csv')
person_data.set_index('id', inplace=True) # set 'id' as primary key
print(person_data.head())  # Print the first few rows to check

# import the person_name_race.csv
# bring in all columns
person_race_name = pd.read_csv('./data/person_name_race.csv')
person_race_name.set_index('id', inplace=True) # set 'id' as primary key
print(person_race_name.head())  # Print the first few rows to check

# import the person_name_race.csv
# bring in 'id', 'name', 'image_race' columns
person_race_image = pd.read_csv('./data/person_tmdb_image_race.csv', usecols=['id', 'name', 'image_race'])
person_race_image.set_index('id', inplace=True) # set 'id' as primary key
print(person_race_image.head())  # Print the first few rows to check


import sqlite3

# Establish the database connection
conn = sqlite3.connect('movies_db.sqlite')

# Write the dataframes to tables in the SQLite database
movie_data.to_sql('movie_data', conn, if_exists='replace')
movie_to_persons.to_sql('movie_to_persons', conn, if_exists='replace')
person_data.to_sql('person_data', conn, if_exists='replace')
person_race_name.to_sql('person_race_name', conn, if_exists='replace')
person_race_image.to_sql('person_race_image', conn, if_exists='replace')

# Close the database connection
conn.close()
